비기능: blackjackApi 인터페이스화
